requirements.txt
# Wheel 0.25+ needed to install certain packages on CPython 3.5+
# like Pillow and psycopg2
# See http://bitly.com/wheel-building-fails-CPython-35
# Verified bug on Python 3.5.1
wheel==0.29.0

# Bleeding edge Django
django==1.10.5

# Configuration
django-environ==0.4.1
whitenoise==3.2.3


# Forms
django-braces==1.10.0
django-crispy-forms==1.6.1

# Models
django-model-utils==2.6.1

# Images
Pillow==4.0.0

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.30.0


# Python-PostgreSQL Database Adapter
psycopg2==2.6.2

# Unicode slugification
awesome-slugify==1.6.5

# Time zones support
pytz==2016.10

# Redis support
django-redis==4.7.0
redis>=2.10.5



coverage==4.3.1
django-coverage-plugin==1.3.1
Sphinx==1.5.1
django-extensions==1.7.5
Werkzeug==0.11.15
django-test-plus==1.0.16
factory-boy==2.8.1

django-debug-toolbar==1.6

# improved REPL
ipdb==0.10.1

pytest-django==3.1.2
pytest-sugar==0.8.0


# Pro-tip: Try not to put anything here. Avoid dependencies in
#	production that aren't in development.




# WSGI Handler
# ------------------------------------------------
gevent==1.2.1
gunicorn==19.6.0

# Static and Media Storage
# ------------------------------------------------
boto==2.45.0
django-storages-redux==1.3.2


# Email backends for Mailgun, Postmark, SendGrid and more
# -------------------------------------------------------
django-anymail==0.7

# Raven is the Sentry client
# --------------------------
raven==5.32.0



flake8==3.2.1 # pyup: != 2.6.0


amqp>=1.4.9
anyjson>=0.3.3
billiard>=3.3.0.23
colorama>=0.3.7
django-bootstrap3-datetimepicker>=2.2.3
django-excel==0.0.4
django-filter>=0.13.0
django-multiupload==0.5.1
django-password-reset==0.9
django-suit>=0.2.19
djangorestframework>=3.4.7
et-xmlfile>=1.0.1
ezodf>=0.3.2
future>=0.15.2
jdcal>=1.3
kombu>=3.0.35
lxml>=3.6.4
odfpy>=0.9.6
openpyxl>=2.4.0
pyexcel>=0.2.5
pyexcel-io>=0.2.3
pyexcel-ods>=0.2.0
pyexcel-ods3>=0.2.1
pyexcel-webio>=0.0.7
pyexcel-xlsx>=0.2.2
texttable==0.8.4
vine==1.1.1
xlrd>=1.0.0
